+++ src/pages/admin/FAQ/index.tsx
@@ -1,61 +1,40 @@
 import { useState } from 'react';
 import Head from 'next/head';
-import { Trash2 } from 'lucide-react';
 import { useCookies } from 'react-cookie';
-import { deleteFaq } from '@/apis';
+import { toast } from 'react-hot-toast';
 import Heading from '@/components/common/Heading';
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger,
-} from '@/components/ui/accordion';
+import FAQList from '@/components/faq/FAQList';
 import { useAllFaq } from '@/hooks/api/faq/useAllFaq';
 import { useCreateFaq } from '@/hooks/api/faq/useCreateFaq';
-
-interface FAQItem {
-  question: string;
-  reply: string;
-}
 
 export default function Index() {
   const [cookies] = useCookies(['token', 'role']);
   const { token } = cookies;
-
-  const { data: FAQ, isLoading, error } = useAllFaq(token);
+  const { data: FAQ } = useAllFaq(token);
   const { mutate: createFaq, isLoading: isSaving } = useCreateFaq();
-  console.log(FAQ?.data, 'FAQ');
 
   const [isEditing, setIsEditing] = useState(false);
-  const [newFAQs, setNewFAQs] = useState<FAQItem[]>([]);
-
-  const safeFAQ: FAQItem[] = Array.isArray(FAQ?.data) ? FAQ?.data : [];
+  const [newFAQs, setNewFAQs] = useState<{ question: string; reply: string }[]>(
+    [],
+  );
 
   const addFAQ = () => {
-    const newFAQ: FAQItem = {
-      question: '질문을 입력해주세요',
-      reply: '답변을 입력해주세요',
-    };
-    setNewFAQs([...newFAQs, newFAQ]);
+    setNewFAQs([
+      ...newFAQs,
+      { question: '질문을 입력해주세요', reply: '답변을 입력해주세요' },
+    ]);
   };
 
-  const saveFAQ = () => {
-    if (newFAQs.length === 0) {
-      return;
+  const saveFAQ = async () => {
+    if (newFAQs.length === 0) return;
+
+    try {
+      await Promise.all(newFAQs.map((faq) => createFaq({ token, ...faq })));
+      toast.success('FAQ가 성공적으로 저장되었습니다');
+      setNewFAQs([]);
+    } catch (error) {
+      toast.error('FAQ 저장에 실패하였습니다');
     }
-
-    newFAQs.forEach((faq) => {
-      if (!faq.question || !faq.reply) {
-        alert('질문과 답변 모두 입력부탁');
-        return;
-      }
-
-      createFaq({ token, ...faq });
-    });
-  };
-
-  const isClickedDeleteButton = (questionId: number) => {
-    deleteFaq({ questionId, token });
   };
 
   return (
@@ -63,6 +42,7 @@ export default function Index() {
       <Head>
         <title>띵동 - FAQ</title>
       </Head>
+
       <div className="flex items-end justify-between">
         <Heading>FAQ 관리</Heading>
         {isEditing ? (
@@ -75,10 +55,13 @@ export default function Index() {
             </button>
             <button
               onClick={saveFAQ}
-              className="ml-3 h-10 rounded-lg bg-blue-500 px-4.5 py-2 text-sm font-bold text-white hover:bg-blue-600"
+              className={`ml-3 h-10 rounded-lg px-4.5 py-2 text-sm font-bold text-white 
+${
+  isSaving ? 'cursor-not-allowed bg-gray-500' : 'bg-blue-500 hover:bg-blue-600'
+}`}
               disabled={isSaving}
             >
-              {isSaving ? '저장 중...' : '저장하기'}
+              저장하기
             </button>
           </div>
         ) : (
@@ -90,8 +73,9 @@ export default function Index() {
           </button>
         )}
       </div>
+
       {isEditing && (
-        <div className="flex w-full justify-end pt-10">
+        <div className="flex w-full flex-row justify-end pt-6">
           <div
             onClick={addFAQ}
             className="flex w-16 cursor-pointer justify-end border-b-2 border-gray-600 pb-0 font-bold text-gray-600"
@@ -100,117 +84,13 @@ export default function Index() {
           </div>
         </div>
       )}
-      <div className="mt-10 flex flex-col items-center justify-center">
-        {safeFAQ.length === 0 && newFAQs.length === 0 && (
-          <div className="text-gray-500">등록된 게시물이 없습니다</div>
-        )}
-        {isEditing ? (
-          <>
-            {newFAQs.map((item, index) => (
-              <Accordion
-                key={index}
-                type="single"
-                collapsible
-                className="w-full"
-              >
-                <AccordionItem value={`item-${index}`}>
-                  <AccordionTrigger isArrow={false}>
-                    <div>
-                      <span>Q</span>
-                      <span className="text-blue-500">.</span>
-                      <span>&nbsp;&nbsp;&nbsp;&nbsp;</span>
-                    </div>
 
-                    <input
-                      className="mx-0 w-11/12 pr-10"
-                      placeholder={item.question}
-                      value={item.question}
-                      onChange={(e) =>
-                        setNewFAQs((prev) =>
-                          prev.map((faq, i) =>
-                            i === index
-                              ? { ...faq, question: e.target.value }
-                              : faq,
-                          ),
-                        )
-                      }
-                    />
-                    <Trash2
-                      className=" text-red-400"
-                      onClick={() => isClickedDeleteButton(item.id)}
-                    />
-                  </AccordionTrigger>
-                  <AccordionContent>
-                    <div className="font-bold text-blue-500">A.</div>
-                    <span>&nbsp;&nbsp;&nbsp;&nbsp;</span>
-                    <input
-                      placeholder={item.reply}
-                      className="w-full bg-gray-50"
-                      value={item.reply}
-                      onChange={(e) =>
-                        setNewFAQs((prev) =>
-                          prev.map((faq, i) =>
-                            i === index
-                              ? { ...faq, reply: e.target.value }
-                              : faq,
-                          ),
-                        )
-                      }
-                    />
-                  </AccordionContent>
-                </AccordionItem>
-              </Accordion>
-            ))}
-            {safeFAQ.map((item, index) => (
-              <Accordion
-                key={index}
-                type="single"
-                collapsible
-                className="w-full"
-              >
-                <AccordionItem value={`item-${index}`}>
-                  <AccordionTrigger isArrow={false}>
-                    <div>
-                      <span>Q</span>
-                      <span className="text-blue-500">.</span>
-                      <span>&nbsp;&nbsp;&nbsp;&nbsp;</span>
-                      <span className="text-left">{item.question}</span>
-                    </div>
-
-                    <Trash2
-                      className="text-red-400"
-                      onClick={() => isClickedDeleteButton(item.id)}
-                    />
-                  </AccordionTrigger>
-                  <AccordionContent>
-                    <div className="font-bold text-blue-500">A.</div>
-                    <span>&nbsp;&nbsp;&nbsp;&nbsp;</span>
-                    {item.reply}
-                  </AccordionContent>
-                </AccordionItem>
-              </Accordion>
-            ))}
-          </>
-        ) : (
-          safeFAQ.map((item, index) => (
-            <Accordion key={index} type="single" collapsible className="w-full">
-              <AccordionItem value={`item-${index}`}>
-                <AccordionTrigger>
-                  <span>Q</span>
-                  <span className="text-blue-500">.</span>
-                  <span>&nbsp;&nbsp;&nbsp;&nbsp;</span>
-                  <span className="text-left">{item.question}</span>
-                </AccordionTrigger>
-                <AccordionContent>
-                  <div className="font-bold text-blue-500">A.</div>
-                  <span>&nbsp;&nbsp;&nbsp;&nbsp;</span>
-                  {item.reply}
-                </AccordionContent>
-              </AccordionItem>
-            </Accordion>
-          ))
-        )}
-      </div>
+      <FAQList
+        FAQ={FAQ}
+        newFAQs={newFAQs}
+        setNewFAQs={setNewFAQs}
+        isEditing={isEditing}
+      />
     </>
   );
 }
